refactor: 댓글 입력창 관심사 분리
