--- src/constants/inputType.js
@@ -1,4 +1,0 @@
-export const INPUT_TYPE = {
-  CHAT: 'chat',
-  POST: 'post',
-};
